Add sms vote validation (#508)
This PR closes #506 .

Plips now may require a SMS confirmation. After signing a plip, if it requires a confirmation, we redirect the user to the phone confirmation screen. The user can't skip this screen, but after entering the phone number, the code confirmation screen is presented, which allows the user to skip this process altogether.

- Feature
Added SMS vote validation