Upgrade Facebook sdk (#516)
This PR closes #515 .

Facebook SDK needed to be upgraded and in order to do so, things needed to happen:

Migrate project to AndroiX (jetfied)
iOS no longer uses cocoapods to install de FB SDK. There were several issues when trying to upgrade.
Android is now using the latest SDK 7.1
iOS is not using the latest SDK because we can't upgrade to the latest react-native-fbsdk. SDK used is 5.15.1.

Feature
added new install script which needs to be run once yarn install-facebook-ios-sdk. This will install the iOS Facebook SDK on your ~/Documents/FacebookSDK folder
Upgrade cocoapods to 1.6.2 (not really necessary after all the install above, but it was done).