fix(dynamic-link): [ISSUE-533] Fix accepting dynamic link in ios (#534)
This PR closes #533 .

When the app is not in memory (has not been started previously), accepting a dynamic link would only open the app, but there would be no link to handle. This was because didFinishLaunchingWithOptions was returning whatever the facebook sdk return when it handled the startup. When this method returns @NO, the app does not receive link callbacks (application:(UIApplication *)application openURL:(NSURL *)url ...) and we couldn't process any link.

This PR also adds a feature to continue to the message sign when following a link and creating a new profile.

Fix
fix(dynamic-link): [ISSUE-533] Fix accepting dynamic link in ios
Feat
go back to message sign after following a link and creating a new profile